+++ examples/run.rs
@@ -1,7 +1,7 @@
 use kaylang::{
     compiler::{artifacts::Artifacts, Compiler},
     src_file::SrcFile,
-    syntax::{ast::Ast, tokenizer::Tokenizer},
+    syntax::{ast::Parser, tokenizer::Tokenizer},
     Color, Logger, ASSEMBLING, ASSEMBLING_ERROR, BUILDING_AST, CHECKING, COMPILING,
     COULD_NOT_RUN_ASSEMBLER, COULD_NOT_RUN_EXECUTABLE, COULD_NOT_RUN_LINKER, GENERATING_ASM,
     LINKING, LINKING_ERROR, LOADING_SOURCE, RUNNING, SUBSTEP_DONE, TOKENIZATION,
@@ -54,7 +54,7 @@ fn main() -> ExitCode {
 
     let ast = {
         let building_ast_sub_step = Logger::new();
-        let building_ast_result = Ast::build(&src, &tokens);
+        let building_ast_result = Parser::parse(&src, &tokens);
         building_ast_sub_step.sub_step_done(&BUILDING_AST);
         match building_ast_result {
             Ok(ast) => ast,